test(geScopeParsingResult): move dir
