+++ src/test/suite/cairo-parser/getScopeParsingResult/storage_var.test.ts
@@ -1,49 +1,47 @@
 import * as assert from "assert";
 import * as path from "path";
 import * as fs from "fs";
-import CairoParser from "../../../lib/CairoParser";
-import FunctionCommentDescParser from "../../../lib/parser/function-comment-new/desc";
-import FunctionSignatureRegexParser from "../../../lib/parser/function-signature/regex";
-import FunctionCommentImplicitArgsParser from "../../../lib/parser/function-comment-new/implicit-args";
-import FunctionCommentExplicitArgsParser from "../../../lib/parser/function-comment-new/explicit-args";
-import FunctionCommentReturnsParser from "../../../lib/parser/function-comment-new/returns";
-import FunctionCommentRaisesParser from "../../../lib/parser/function-comment-new/raises";
+import CairoParser from "../../../../lib/CairoParser";
+import FunctionCommentDescParser from "../../../../lib/parser/function-comment-new/desc";
+import FunctionSignatureRegexParser from "../../../../lib/parser/function-signature/regex";
+import FunctionCommentImplicitArgsParser from "../../../../lib/parser/function-comment-new/implicit-args";
+import FunctionCommentExplicitArgsParser from "../../../../lib/parser/function-comment-new/explicit-args";
+import FunctionCommentReturnsParser from "../../../../lib/parser/function-comment-new/returns";
+import FunctionCommentRaisesParser from "../../../../lib/parser/function-comment-new/raises";
 
-suite("getScopeParsingResult: namespace", () => {
-  test("should get `5` for the length of function scope", () => {
+suite("getScopeParsingResult: storage_var", () => {
+  test("should get `6` for the length of function scope", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testContracts/ERC20Compliant/library.cairo"
+      "../../../../../testContracts/ERC20Compliant/library.cairo"
     );
     const text = fs.readFileSync(pathFile, "utf8");
-    // parse whole scope
-    const functionScopes = CairoParser.parseNamespaceScopes(text);
-    assert.equal(functionScopes!.length, 5, "failed to parse");
 
-    const resultScope = CairoParser.getScopeParsingResult(text, "namespace");
-    assert.equal(resultScope!.length, 5, "failed to parse");
+    // parse whole scope
+    const functionScopes = CairoParser.parseFunctionScope(text, "storage_var");
+    assert.equal(functionScopes!.length, 6);
+
+    const resultScope = CairoParser.getScopeParsingResult(text, "storage_var");
+    assert.equal(resultScope!.length, 6);
   });
-
-  test("should get `constructor` function scope", () => {
+  test("should get `ERC20_name` function scope", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testContracts/ERC20Namespace/library.cairo"
+      "../../../../../testContracts/ERC20Compliant/library.cairo"
     );
     const text = fs.readFileSync(pathFile, "utf8");
     // parse whole scope
-    const functionScopes = CairoParser.parseNamespaceScopes(text);
+    const functionScopes = CairoParser.parseFunctionScope(text, "storage_var");
 
     // Function signature parsing
     const functionSignatureParser = new FunctionSignatureRegexParser();
 
     // Comment parsing
     // parse comment lines
+
     const scopeNumber = 0;
-    const functionScope = functionScopes![scopeNumber];
-
     const functionCommentScope = CairoParser.parseCommentLines(
-      functionScopes![scopeNumber],
-      true
+      functionScopes![scopeNumber]
     )!;
 
     const functionCommentText: string = functionCommentScope!.text.join("");
@@ -64,245 +62,12 @@ suite("getScopeParsingResult: namespace", () => {
 
     const parsingTarget = [
       {
-        attributeName: "namespace ERC20",
-        functionName: "constructor",
+        attributeName: "storage_var",
+        functionName: "ERC20_name",
         functionSignature: {
-          implicitArgs: [
-            { name: "syscall_ptr", type: "felt*" },
-            { name: "pedersen_ptr", type: "HashBuiltin*" },
-            { name: "range_check_ptr", type: "" },
-          ],
-          explicitArgs: [
-            { name: "name", type: "felt" },
-            { name: "symbol", type: "felt" },
-            { name: "multiplier", type: "felt" },
-          ],
-          returns: null,
-        },
-        functionComment: {
-          desc: [
-            {
-              name: "",
-              type: "",
-              desc: "Initializes the contract with the given name, symbol, and decimals",
-              charIndex: { start: 29, end: 95 },
-            },
-          ],
-          implicitArgs: [
-            {
-              name: "syscall_ptr",
-              type: "felt*",
-              desc: "",
-              charIndex: { start: 133, end: 151 },
-            },
-            {
-              name: "pedersen_ptr",
-              type: "HashBuiltin*",
-              desc: "",
-              charIndex: { start: 164, end: 190 },
-            },
-            {
-              name: "range_check_ptr",
-              type: "",
-              desc: "",
-              charIndex: { start: 203, end: 218 },
-            },
-          ],
-          explicitArgs: [
-            {
-              name: "name",
-              type: "felt",
-              desc: "The name of the token",
-              charIndex: { start: 256, end: 289 },
-            },
-            {
-              name: "symbol",
-              type: "felt",
-              desc: "The symbol of the token",
-              charIndex: { start: 302, end: 339 },
-            },
-            {
-              name: "multiplier",
-              type: "felt",
-              desc: "The multiplier of the token",
-              charIndex: { start: 352, end: 397 },
-            },
-          ],
-          returns: null,
-          raises: null,
-          charIndex: { start: 2992, end: 3389 },
-        },
-      },
-    ];
-
-    const parsingOutput = [
-      {
-        attributeName: functionSignatureParser.getAttributeName(
-          functionScopes![scopeNumber].text
-        ),
-        functionName: functionSignatureParser.getFunctionName(
-          functionScopes![scopeNumber].text
-        ),
-        functionSignature: {
-          implicitArgs: functionSignatureParser.getImplicitArgs(
-            functionScopes![scopeNumber].text
-          ),
-          explicitArgs: functionSignatureParser.getExplicitArgs(
-            functionScopes![scopeNumber].text
-          ),
-          returns: functionSignatureParser.getReturns(
-            functionScopes![scopeNumber].text
-          ),
-        },
-        functionComment: {
-          desc: functionCommentDescParser.parseCommentLines(
-            functionCommentScope!.text
-          ),
-          implicitArgs: functionCommentImplicitArgsParser.parseCommentLines(
-            functionCommentScope!.text
-          ),
-          explicitArgs: functionCommentExplicitArgsParser.parseCommentLines(
-            functionCommentScope!.text
-          ),
-          returns: functionCommentReturnsParser.parseCommentLines(
-            functionCommentScope!.text
-          ),
-          raises: functionCommentRaisesParser.parseCommentLines(
-            functionCommentScope!.text
-          ),
-          charIndex: {
-            start: functionCommentScope!.start,
-            end: functionCommentScope!.end,
-          },
-        },
-      },
-    ];
-
-    assert.deepEqual(parsingTarget, parsingOutput, "failed to parse");
-
-    const parsingResult = CairoParser.getScopeParsingResult(text, "namespace")![
-      scopeNumber
-    ];
-    assert.deepEqual(parsingResult, parsingOutput[0], "failed to parse");
-
-    const { charIndex, ...parsingOutputWithoutCharIndex } =
-      parsingOutput[0].functionComment;
-
-    var commentParsingResult = [];
-
-    for (let [key, values] of Object.entries(parsingOutputWithoutCharIndex)) {
-      if (values) {
-        for (const value of values) {
-          const charIndex = value.charIndex;
-          var char = "";
-          for (
-            let i = functionCommentScope!.start + charIndex.start;
-            i < functionCommentScope!.start + charIndex.end;
-            i++
-          ) {
-            char += text.at(i);
-          }
-          const commentParsing = {
-            [key]: char,
-          };
-          commentParsingResult.push(commentParsing);
-        }
-      }
-    }
-    const textTarget = [
-      {
-        desc: "Initializes the contract with the given name, symbol, and decimals",
-      },
-      { implicitArgs: "syscall_ptr(felt*)" },
-      { implicitArgs: "pedersen_ptr(HashBuiltin*)" },
-      { implicitArgs: "range_check_ptr" },
-      { explicitArgs: "name(felt): The name of the token" },
-      {
-        explicitArgs: "symbol(felt): The symbol of the token",
-      },
-      {
-        explicitArgs: "multiplier(felt): The multiplier of the token",
-      },
-    ];
-    assert.deepEqual(textTarget, commentParsingResult, "failed to parse");
-
-    var functionCommentParsingResult = "";
-    for (var i = charIndex!.start; i < charIndex!.end; i++) {
-      functionCommentParsingResult += text.at(i);
-    }
-
-    const functionCommentTarget = `
-        # Desc:
-        #   Initializes the contract with the given name, symbol, and decimals
-        # Implicit args:
-        #   syscall_ptr(felt*)
-        #   pedersen_ptr(HashBuiltin*)
-        #   range_check_ptr
-        # Explicit args:
-        #   name(felt): The name of the token
-        #   symbol(felt): The symbol of the token
-        #   multiplier(felt): The multiplier of the token`;
-
-    assert.equal(
-      functionCommentTarget,
-      functionCommentParsingResult,
-      "failed to parse"
-    );
-  });
-
-  test("should get `name` function scope", () => {
-    const pathFile = path.resolve(
-      __dirname,
-      "../../../../testContracts/ERC20Namespace/library.cairo"
-    );
-    const text = fs.readFileSync(pathFile, "utf8");
-    // parse whole scope
-    const functionScopes = CairoParser.parseNamespaceScopes(text);
-
-    // Function signature parsing
-    const functionSignatureParser = new FunctionSignatureRegexParser();
-
-    // Comment parsing
-    // parse comment lines
-    const scopeNumber = 1;
-    const functionCommentScope = CairoParser.parseCommentLines(
-      functionScopes![scopeNumber],
-      true
-    )!;
-
-    const functionCommentText: string = functionCommentScope!.text.join("");
-
-    const functionCommentDescParser = new FunctionCommentDescParser(
-      functionCommentText
-    );
-    const functionCommentImplicitArgsParser =
-      new FunctionCommentImplicitArgsParser(functionCommentText);
-    const functionCommentExplicitArgsParser =
-      new FunctionCommentExplicitArgsParser(functionCommentText);
-    const functionCommentReturnsParser = new FunctionCommentReturnsParser(
-      functionCommentText
-    );
-    const functionCommentRaisesParser = new FunctionCommentRaisesParser(
-      functionCommentText
-    );
-
-    const parsingTarget = [
-      {
-        attributeName: "namespace ERC20",
-        functionName: "name",
-        functionSignature: {
-          implicitArgs: [
-            { name: "syscall_ptr", type: "felt*" },
-            { name: "pedersen_ptr", type: "HashBuiltin*" },
-            { name: "range_check_ptr", type: "" },
-          ],
+          implicitArgs: null,
           explicitArgs: null,
-          returns: [
-            {
-              name: "name",
-              type: "felt",
-            },
-          ],
+          returns: [{ name: "name", type: "felt" }],
         },
         functionComment: {
           desc: [
@@ -310,62 +75,41 @@ suite("getScopeParsingResult: namespace", () => {
               name: "",
               type: "",
               desc: "Returns the name of the token",
-              charIndex: { start: 29, end: 58 },
+              charIndex: { start: 22, end: 51 },
             },
           ],
-          implicitArgs: [
-            {
-              name: "syscall_ptr",
-              type: "felt*",
-              desc: "",
-              charIndex: { start: 96, end: 114 },
-            },
-            {
-              name: "pedersen_ptr",
-              type: "HashBuiltin*",
-              desc: "",
-              charIndex: { start: 127, end: 153 },
-            },
-            {
-              name: "range_check_ptr",
-              type: "",
-              desc: "",
-              charIndex: { start: 166, end: 181 },
-            },
-          ],
+          implicitArgs: null,
           explicitArgs: null,
           returns: [
             {
               name: "name",
               type: "felt",
               desc: "The name of the token",
-              charIndex: { start: 213, end: 246 },
+              charIndex: { start: 75, end: 108 },
             },
           ],
           raises: null,
-          charIndex: { start: 3782, end: 4028 },
+          charIndex: { start: 1306, end: 1414 },
         },
       },
     ];
 
-    const parsingOutput = [
+    var parsingOutput = [
       {
         attributeName: functionSignatureParser.getAttributeName(
-          functionScopes![scopeNumber].text
+          functionScopes![0].text
         ),
         functionName: functionSignatureParser.getFunctionName(
-          functionScopes![scopeNumber].text
+          functionScopes![0].text
         ),
         functionSignature: {
           implicitArgs: functionSignatureParser.getImplicitArgs(
-            functionScopes![scopeNumber].text
+            functionScopes![0].text
           ),
           explicitArgs: functionSignatureParser.getExplicitArgs(
-            functionScopes![scopeNumber].text
+            functionScopes![0].text
           ),
-          returns: functionSignatureParser.getReturns(
-            functionScopes![scopeNumber].text
-          ),
+          returns: functionSignatureParser.getReturns(functionScopes![0].text),
         },
         functionComment: {
           desc: functionCommentDescParser.parseCommentLines(
@@ -393,9 +137,10 @@ suite("getScopeParsingResult: namespace", () => {
 
     assert.deepEqual(parsingTarget, parsingOutput, "failed to parse");
 
-    const parsingResult = CairoParser.getScopeParsingResult(text, "namespace")![
-      scopeNumber
-    ];
+    const parsingResult = CairoParser.getScopeParsingResult(
+      text,
+      "storage_var"
+    )![scopeNumber];
     assert.deepEqual(parsingResult, parsingOutput[0], "failed to parse");
 
     const { charIndex, ...parsingOutputWithoutCharIndex } =
@@ -423,12 +168,7 @@ suite("getScopeParsingResult: namespace", () => {
       }
     }
     const textTarget = [
-      {
-        desc: "Returns the name of the token",
-      },
-      { implicitArgs: "syscall_ptr(felt*)" },
-      { implicitArgs: "pedersen_ptr(HashBuiltin*)" },
-      { implicitArgs: "range_check_ptr" },
+      { desc: "Returns the name of the token" },
       { returns: "name(felt): The name of the token" },
     ];
     assert.deepEqual(textTarget, commentParsingResult, "failed to parse");
@@ -439,14 +179,10 @@ suite("getScopeParsingResult: namespace", () => {
     }
 
     const functionCommentTarget = `
-        # Desc:
-        #   Returns the name of the token
-        # Implicit args:
-        #   syscall_ptr(felt*)
-        #   pedersen_ptr(HashBuiltin*)
-        #   range_check_ptr
-        # Returns:
-        #   name(felt): The name of the token`;
+    # Desc: 
+    #   Returns the name of the token
+    # Returns:
+    #   name(felt): The name of the token`;
 
     assert.equal(
       functionCommentTarget,
@@ -455,14 +191,180 @@ suite("getScopeParsingResult: namespace", () => {
     );
   });
 
-  test("should get `transfer_from` function scope", () => {
+  test("should get `ERC20_symbol` function scope", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testContracts/ERC20Namespace/library.cairo"
+      "../../../../../testContracts/ERC20Compliant/library.cairo"
     );
     const text = fs.readFileSync(pathFile, "utf8");
+
     // parse whole scope
-    const functionScopes = CairoParser.parseNamespaceScopes(text);
+    const functionScopes = CairoParser.parseFunctionScope(text, "storage_var");
+
+    // Function signature parsing
+    const functionSignatureParser = new FunctionSignatureRegexParser();
+
+    // Comment parsing
+    // parse comment lines
+    const scopeNumber = 1;
+    const functionCommentScope = CairoParser.parseCommentLines(
+      functionScopes![scopeNumber]
+    )!;
+
+    const functionCommentText: string = functionCommentScope!.text.join("");
+
+    const functionCommentDescParser = new FunctionCommentDescParser(
+      functionCommentText
+    );
+    const functionCommentImplicitArgsParser =
+      new FunctionCommentImplicitArgsParser(functionCommentText);
+    const functionCommentExplicitArgsParser =
+      new FunctionCommentExplicitArgsParser(functionCommentText);
+    const functionCommentReturnsParser = new FunctionCommentReturnsParser(
+      functionCommentText
+    );
+    const functionCommentRaisesParser = new FunctionCommentRaisesParser(
+      functionCommentText
+    );
+
+    const parsingTarget = [
+      {
+        attributeName: "storage_var",
+        functionName: "ERC20_symbol",
+        functionSignature: {
+          implicitArgs: null,
+          explicitArgs: null,
+          returns: [{ name: "symbol", type: "felt" }],
+        },
+        functionComment: {
+          desc: [
+            {
+              name: "",
+              type: "",
+              desc: "Returns the symbol of the token",
+              charIndex: { start: 22, end: 53 },
+            },
+          ],
+          implicitArgs: null,
+          explicitArgs: null,
+          returns: [
+            {
+              name: "symbol",
+              type: "felt",
+              desc: "The symbol of the token",
+              charIndex: { start: 77, end: 114 },
+            },
+          ],
+          raises: null,
+          charIndex: { start: 1472, end: 1586 },
+        },
+      },
+    ];
+
+    var parsingOutput = [
+      {
+        attributeName: functionSignatureParser.getAttributeName(
+          functionScopes![scopeNumber].text
+        ),
+        functionName: functionSignatureParser.getFunctionName(
+          functionScopes![scopeNumber].text
+        ),
+        functionSignature: {
+          implicitArgs: functionSignatureParser.getImplicitArgs(
+            functionScopes![scopeNumber].text
+          ),
+          explicitArgs: functionSignatureParser.getExplicitArgs(
+            functionScopes![scopeNumber].text
+          ),
+          returns: functionSignatureParser.getReturns(
+            functionScopes![scopeNumber].text
+          ),
+        },
+        functionComment: {
+          desc: functionCommentDescParser.parseCommentLines(
+            functionCommentScope!.text
+          ),
+          implicitArgs: functionCommentImplicitArgsParser.parseCommentLines(
+            functionCommentScope!.text
+          ),
+          explicitArgs: functionCommentExplicitArgsParser.parseCommentLines(
+            functionCommentScope!.text
+          ),
+          returns: functionCommentReturnsParser.parseCommentLines(
+            functionCommentScope!.text
+          ),
+          raises: functionCommentRaisesParser.parseCommentLines(
+            functionCommentScope!.text
+          ),
+          charIndex: {
+            start: functionCommentScope!.start,
+            end: functionCommentScope!.end,
+          },
+        },
+      },
+    ];
+
+    assert.deepEqual(parsingTarget, parsingOutput, "failed to parse");
+
+    const { charIndex, ...parsingOutputWithoutCharIndex } =
+      parsingOutput[0].functionComment;
+
+    assert.deepEqual(parsingTarget, parsingOutput, "failed to parse");
+
+    var commentParsingResult = [];
+
+    for (let [key, values] of Object.entries(parsingOutputWithoutCharIndex)) {
+      if (values) {
+        for (const value of values) {
+          const charIndex = value.charIndex;
+          var char = "";
+          for (
+            let i = functionCommentScope!.start + charIndex.start;
+            i < functionCommentScope!.start + charIndex.end;
+            i++
+          ) {
+            char += text.at(i);
+          }
+          const commentParsing = {
+            [key]: char,
+          };
+          commentParsingResult.push(commentParsing);
+        }
+      }
+    }
+    const textTarget = [
+      { desc: "Returns the symbol of the token" },
+      { returns: "symbol(felt): The symbol of the token" },
+    ];
+    assert.deepEqual(textTarget, commentParsingResult, "failed to parse");
+
+    var functionCommentParsingResult = "";
+    for (var i = charIndex!.start; i < charIndex!.end; i++) {
+      functionCommentParsingResult += text.at(i);
+    }
+
+    const functionCommentTarget = `
+    # Desc: 
+    #   Returns the symbol of the token
+    # Returns:
+    #   symbol(felt): The symbol of the token`;
+
+    assert.equal(
+      functionCommentTarget,
+      functionCommentParsingResult,
+      "failed to parse"
+    );
+  });
+
+  test("should get `ERC20_decimals` function scope", () => {
+    const pathFile = path.resolve(
+      __dirname,
+      "../../../../../testContracts/ERC20Compliant/library.cairo"
+    );
+    const text = fs.readFileSync(pathFile, "utf8");
+
+    // parse whole scope
+    const functionScopes = CairoParser.parseFunctionScope(text, "storage_var");
 
     // Function signature parsing
     const functionSignatureParser = new FunctionSignatureRegexParser();
@@ -471,9 +373,9 @@ suite("getScopeParsingResult: namespace", () => {
     // parse comment lines
     const scopeNumber = 2;
     const functionCommentScope = CairoParser.parseCommentLines(
-      functionScopes![scopeNumber],
-      true
+      functionScopes![scopeNumber]
     )!;
+
     const functionCommentText: string = functionCommentScope!.text.join("");
 
     const functionCommentDescParser = new FunctionCommentDescParser(
@@ -492,78 +394,39 @@ suite("getScopeParsingResult: namespace", () => {
 
     const parsingTarget = [
       {
-        attributeName: "namespace ERC20",
-        functionName: "transfer_from",
+        attributeName: "storage_var",
+        functionName: "ERC20_decimals",
         functionSignature: {
-          implicitArgs: [
-            { name: "syscall_ptr", type: "felt*" },
-            { name: "pedersen_ptr", type: "HashBuiltin*" },
-            { name: "range_check_ptr", type: "" },
-          ],
-          explicitArgs: [
-            { name: "sender", type: "felt" },
-            { name: "recipient", type: "felt" },
-            { name: "amount", type: "Uint256" },
-          ],
-          returns: null,
+          implicitArgs: null,
+          explicitArgs: null,
+          returns: [{ name: "decimals", type: "Uint256" }],
         },
         functionComment: {
           desc: [
             {
               name: "",
               type: "",
-              desc: "Transfers tokens from one account to another",
-              charIndex: { start: 29, end: 73 },
+              desc: "Returns the number of decimals of the token",
+              charIndex: { start: 22, end: 65 },
             },
           ],
-          implicitArgs: [
+          implicitArgs: null,
+          explicitArgs: null,
+          returns: [
             {
-              name: "syscall_ptr",
-              type: "felt*",
-              desc: "",
-              charIndex: { start: 111, end: 129 },
-            },
-            {
-              name: "pedersen_ptr",
-              type: "HashBuiltin*",
-              desc: "",
-              charIndex: { start: 142, end: 168 },
-            },
-            {
-              name: "range_check_ptr",
-              type: "",
-              desc: "",
-              charIndex: { start: 181, end: 196 },
-            },
-          ],
-          explicitArgs: [
-            {
-              name: "sender",
-              type: "felt",
-              desc: "The address of the sender",
-              charIndex: { start: 234, end: 273 },
-            },
-            {
-              name: "recipient",
-              type: "felt",
-              desc: "The address of the recipient",
-              charIndex: { start: 286, end: 331 },
-            },
-            {
-              name: "amount",
+              name: "decimals",
               type: "Uint256",
-              desc: "The amount of tokens to be transferred",
-              charIndex: { start: 344, end: 399 },
+              desc: "The number of decimals of the token",
+              charIndex: { start: 89, end: 143 },
             },
           ],
-          returns: null,
           raises: null,
-          charIndex: { start: 4261, end: 4660 },
+          charIndex: { start: 1651, end: 1794 },
         },
       },
     ];
 
-    const parsingOutput = [
+    var parsingOutput = [
       {
         attributeName: functionSignatureParser.getAttributeName(
           functionScopes![scopeNumber].text
@@ -608,13 +471,10 @@ suite("getScopeParsingResult: namespace", () => {
 
     assert.deepEqual(parsingTarget, parsingOutput, "failed to parse");
 
-    const parsingResult = CairoParser.getScopeParsingResult(text, "namespace")![
-      scopeNumber
-    ];
-    assert.deepEqual(parsingResult, parsingOutput[0], "failed to parse");
-
     const { charIndex, ...parsingOutputWithoutCharIndex } =
       parsingOutput[0].functionComment;
+
+    assert.deepEqual(parsingTarget, parsingOutput, "failed to parse");
 
     var commentParsingResult = [];
 
@@ -638,15 +498,8 @@ suite("getScopeParsingResult: namespace", () => {
       }
     }
     const textTarget = [
-      { desc: "Transfers tokens from one account to another" },
-      { implicitArgs: "syscall_ptr(felt*)" },
-      { implicitArgs: "pedersen_ptr(HashBuiltin*)" },
-      { implicitArgs: "range_check_ptr" },
-      { explicitArgs: "sender(felt): The address of the sender" },
-      { explicitArgs: "recipient(felt): The address of the recipient" },
-      {
-        explicitArgs: "amount(Uint256): The amount of tokens to be transferred",
-      },
+      { desc: "Returns the number of decimals of the token" },
+      { returns: "decimals(Uint256): The number of decimals of the token" },
     ];
     assert.deepEqual(textTarget, commentParsingResult, "failed to parse");
 
@@ -656,16 +509,10 @@ suite("getScopeParsingResult: namespace", () => {
     }
 
     const functionCommentTarget = `
-        # Desc:
-        #   Transfers tokens from one account to another
-        # Implicit args:
-        #   syscall_ptr(felt*)
-        #   pedersen_ptr(HashBuiltin*)
-        #   range_check_ptr
-        # Explicit args:
-        #   sender(felt): The address of the sender
-        #   recipient(felt): The address of the recipient
-        #   amount(Uint256): The amount of tokens to be transferred`;
+    # Desc: 
+    #   Returns the number of decimals of the token
+    # Returns:
+    #   decimals(Uint256): The number of decimals of the token`;
 
     assert.equal(
       functionCommentTarget,
@@ -674,14 +521,15 @@ suite("getScopeParsingResult: namespace", () => {
     );
   });
 
-  test("should get `_mint` function scope", () => {
+  test("should get `ERC20_total_supply` function scope", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testContracts/ERC20Namespace/library.cairo"
+      "../../../../../testContracts/ERC20Compliant/library.cairo"
     );
     const text = fs.readFileSync(pathFile, "utf8");
+
     // parse whole scope
-    const functionScopes = CairoParser.parseNamespaceScopes(text);
+    const functionScopes = CairoParser.parseFunctionScope(text, "storage_var");
 
     // Function signature parsing
     const functionSignatureParser = new FunctionSignatureRegexParser();
@@ -690,8 +538,7 @@ suite("getScopeParsingResult: namespace", () => {
     // parse comment lines
     const scopeNumber = 3;
     const functionCommentScope = CairoParser.parseCommentLines(
-      functionScopes![scopeNumber],
-      true
+      functionScopes![scopeNumber]
     )!;
 
     const functionCommentText: string = functionCommentScope!.text.join("");
@@ -712,71 +559,39 @@ suite("getScopeParsingResult: namespace", () => {
 
     const parsingTarget = [
       {
-        attributeName: "namespace internal",
-        functionName: "_mint",
+        attributeName: "storage_var",
+        functionName: "ERC20_total_supply",
         functionSignature: {
-          implicitArgs: [
-            { name: "syscall_ptr", type: "felt*" },
-            { name: "pedersen_ptr", type: "HashBuiltin*" },
-            { name: "range_check_ptr", type: "" },
-          ],
-          explicitArgs: [
-            { name: "recipient", type: "felt" },
-            { name: "amount", type: "Uint256" },
-          ],
-          returns: null,
+          implicitArgs: null,
+          explicitArgs: null,
+          returns: [{ name: "total_supply", type: "Uint256" }],
         },
         functionComment: {
           desc: [
             {
               name: "",
               type: "",
-              desc: "Mints tokens to an account",
-              charIndex: { start: 29, end: 55 },
+              desc: "Returns total amount of tokens in existence",
+              charIndex: { start: 22, end: 65 },
             },
           ],
-          implicitArgs: [
+          implicitArgs: null,
+          explicitArgs: null,
+          returns: [
             {
-              name: "syscall_ptr",
-              type: "felt*",
-              desc: "",
-              charIndex: { start: 93, end: 111 },
-            },
-            {
-              name: "pedersen_ptr",
-              type: "HashBuiltin*",
-              desc: "",
-              charIndex: { start: 124, end: 150 },
-            },
-            {
-              name: "range_check_ptr",
-              type: "",
-              desc: "",
-              charIndex: { start: 163, end: 178 },
-            },
-          ],
-          explicitArgs: [
-            {
-              name: "recipient",
-              type: "felt",
-              desc: "The address of the recipient",
-              charIndex: { start: 216, end: 261 },
-            },
-            {
-              name: "amount",
+              name: "total_supply",
               type: "Uint256",
-              desc: "The amount of tokens to be minted",
-              charIndex: { start: 274, end: 324 },
+              desc: "The total amount of tokens in existence",
+              charIndex: { start: 89, end: 151 },
             },
           ],
-          returns: null,
           raises: null,
-          charIndex: { start: 5071, end: 5431 },
+          charIndex: { start: 1913, end: 2064 },
         },
       },
     ];
 
-    const parsingOutput = [
+    var parsingOutput = [
       {
         attributeName: functionSignatureParser.getAttributeName(
           functionScopes![scopeNumber].text
@@ -820,14 +635,11 @@ suite("getScopeParsingResult: namespace", () => {
     ];
 
     assert.deepEqual(parsingTarget, parsingOutput, "failed to parse");
-
-    const parsingResult = CairoParser.getScopeParsingResult(text, "namespace")![
-      scopeNumber
-    ];
-    assert.deepEqual(parsingResult, parsingOutput[0], "failed to parse");
-
     const { charIndex, ...parsingOutputWithoutCharIndex } =
       parsingOutput[0].functionComment;
+
+    assert.deepEqual(parsingTarget, parsingOutput, "failed to parse");
+
     var commentParsingResult = [];
 
     for (let [key, values] of Object.entries(parsingOutputWithoutCharIndex)) {
@@ -850,12 +662,11 @@ suite("getScopeParsingResult: namespace", () => {
       }
     }
     const textTarget = [
-      { desc: "Mints tokens to an account" },
-      { implicitArgs: "syscall_ptr(felt*)" },
-      { implicitArgs: "pedersen_ptr(HashBuiltin*)" },
-      { implicitArgs: "range_check_ptr" },
-      { explicitArgs: "recipient(felt): The address of the recipient" },
-      { explicitArgs: "amount(Uint256): The amount of tokens to be minted" },
+      { desc: "Returns total amount of tokens in existence" },
+      {
+        returns:
+          "total_supply(Uint256): The total amount of tokens in existence",
+      },
     ];
     assert.deepEqual(textTarget, commentParsingResult, "failed to parse");
 
@@ -865,17 +676,10 @@ suite("getScopeParsingResult: namespace", () => {
     }
 
     const functionCommentTarget = `
-        # Desc:
-        #   Mints tokens to an account
-        # Implicit args:
-        #   syscall_ptr(felt*)
-        #   pedersen_ptr(HashBuiltin*)
-        #   range_check_ptr
-        # Explicit args:
-        #   recipient(felt): The address of the recipient
-        #   amount(Uint256): The amount of tokens to be minted
-        # Returns:
-        #   None`;
+    # Desc: 
+    #   Returns total amount of tokens in existence
+    # Returns:
+    #   total_supply(Uint256): The total amount of tokens in existence`;
 
     assert.equal(
       functionCommentTarget,
@@ -884,14 +688,15 @@ suite("getScopeParsingResult: namespace", () => {
     );
   });
 
-  test("should get `_burn` function scope", () => {
+  test("should get `ERC20_balances` function scope", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testContracts/ERC20Namespace/library.cairo"
+      "../../../../../testContracts/ERC20Compliant/library.cairo"
     );
     const text = fs.readFileSync(pathFile, "utf8");
+
     // parse whole scope
-    const functionScopes = CairoParser.parseNamespaceScopes(text);
+    const functionScopes = CairoParser.parseFunctionScope(text, "storage_var");
 
     // Function signature parsing
     const functionSignatureParser = new FunctionSignatureRegexParser();
@@ -900,8 +705,7 @@ suite("getScopeParsingResult: namespace", () => {
     // parse comment lines
     const scopeNumber = 4;
     const functionCommentScope = CairoParser.parseCommentLines(
-      functionScopes![scopeNumber],
-      true
+      functionScopes![scopeNumber]
     )!;
 
     const functionCommentText: string = functionCommentScope!.text.join("");
@@ -922,71 +726,46 @@ suite("getScopeParsingResult: namespace", () => {
 
     const parsingTarget = [
       {
-        attributeName: "namespace internal",
-        functionName: "_burn",
+        attributeName: "storage_var",
+        functionName: "ERC20_balances",
         functionSignature: {
-          implicitArgs: [
-            { name: "syscall_ptr", type: "felt*" },
-            { name: "pedersen_ptr", type: "HashBuiltin*" },
-            { name: "range_check_ptr", type: "" },
-          ],
-          explicitArgs: [
-            { name: "account", type: "felt" },
-            { name: "amount", type: "Uint256" },
-          ],
-          returns: null,
+          implicitArgs: null,
+          explicitArgs: [{ name: "account", type: "felt" }],
+          returns: [{ name: "balance", type: "Uint256" }],
         },
         functionComment: {
           desc: [
             {
               name: "",
               type: "",
-              desc: "Burns tokens from an account",
-              charIndex: { start: 29, end: 57 },
+              desc: "Returns the amount of tokens owned by an account",
+              charIndex: { start: 22, end: 70 },
             },
           ],
-          implicitArgs: [
-            {
-              name: "syscall_ptr",
-              type: "felt*",
-              desc: "",
-              charIndex: { start: 95, end: 113 },
-            },
-            {
-              name: "pedersen_ptr",
-              type: "HashBuiltin*",
-              desc: "",
-              charIndex: { start: 126, end: 152 },
-            },
-            {
-              name: "range_check_ptr",
-              type: "",
-              desc: "",
-              charIndex: { start: 165, end: 180 },
-            },
-          ],
+          implicitArgs: null,
           explicitArgs: [
             {
               name: "account",
               type: "felt",
-              desc: "The address of the recipient",
-              charIndex: { start: 218, end: 261 },
-            },
-            {
-              name: "amount",
-              type: "Uint256",
-              desc: "The amount of tokens to be burned",
-              charIndex: { start: 274, end: 324 },
+              desc: "The address of the account",
+              charIndex: { start: 100, end: 141 },
             },
           ],
-          returns: null,
+          returns: [
+            {
+              name: "balance",
+              type: "Uint256",
+              desc: "The amount of tokens owned by an account",
+              charIndex: { start: 165, end: 223 },
+            },
+          ],
           raises: null,
-          charIndex: { start: 6474, end: 6798 },
+          charIndex: { start: 2142, end: 2365 },
         },
       },
     ];
 
-    const parsingOutput = [
+    var parsingOutput = [
       {
         attributeName: functionSignatureParser.getAttributeName(
           functionScopes![scopeNumber].text
@@ -1029,10 +808,7 @@ suite("getScopeParsingResult: namespace", () => {
       },
     ];
 
-    const parsingResult = CairoParser.getScopeParsingResult(text, "namespace")![
-      scopeNumber
-    ];
-    assert.deepEqual(parsingResult, parsingOutput[0], "failed to parse");
+    assert.deepEqual(parsingTarget, parsingOutput, "failed to parse");
 
     const { charIndex, ...parsingOutputWithoutCharIndex } =
       parsingOutput[0].functionComment;
@@ -1061,12 +837,9 @@ suite("getScopeParsingResult: namespace", () => {
       }
     }
     const textTarget = [
-      { desc: "Burns tokens from an account" },
-      { implicitArgs: "syscall_ptr(felt*)" },
-      { implicitArgs: "pedersen_ptr(HashBuiltin*)" },
-      { implicitArgs: "range_check_ptr" },
-      { explicitArgs: "account(felt): The address of the recipient" },
-      { explicitArgs: "amount(Uint256): The amount of tokens to be burned" },
+      { desc: "Returns the amount of tokens owned by an account" },
+      { explicitArgs: "account(felt): The address of the account" },
+      { returns: "balance(Uint256): The amount of tokens owned by an account" },
     ];
     assert.deepEqual(textTarget, commentParsingResult, "failed to parse");
 
@@ -1076,15 +849,200 @@ suite("getScopeParsingResult: namespace", () => {
     }
 
     const functionCommentTarget = `
-        # Desc:
-        #   Burns tokens from an account
-        # Implicit args:
-        #   syscall_ptr(felt*)
-        #   pedersen_ptr(HashBuiltin*)
-        #   range_check_ptr
-        # Explicit args:
-        #   account(felt): The address of the recipient
-        #   amount(Uint256): The amount of tokens to be burned`;
+    # Desc: 
+    #   Returns the amount of tokens owned by an account
+    # Explicit args:
+    #   account(felt): The address of the account
+    # Returns:
+    #   balance(Uint256): The amount of tokens owned by an account`;
+
+    assert.equal(
+      functionCommentTarget,
+      functionCommentParsingResult,
+      "failed to parse"
+    );
+  });
+
+  test("should get `ERC20_allowances` function scope", () => {
+    const pathFile = path.resolve(
+      __dirname,
+      "../../../../../testContracts/ERC20Compliant/library.cairo"
+    );
+    const text = fs.readFileSync(pathFile, "utf8");
+
+    // parse whole scope
+    const functionScopes = CairoParser.parseFunctionScope(text, "storage_var");
+
+    // Function signature parsing
+    const functionSignatureParser = new FunctionSignatureRegexParser();
+
+    // Comment parsing
+    // parse comment lines
+    const scopeNumber = 5;
+    const functionCommentScope = CairoParser.parseCommentLines(
+      functionScopes![scopeNumber]
+    )!;
+
+    const functionCommentText: string = functionCommentScope!.text.join("");
+
+    const functionCommentDescParser = new FunctionCommentDescParser(
+      functionCommentText
+    );
+    const functionCommentImplicitArgsParser =
+      new FunctionCommentImplicitArgsParser(functionCommentText);
+    const functionCommentExplicitArgsParser =
+      new FunctionCommentExplicitArgsParser(functionCommentText);
+    const functionCommentReturnsParser = new FunctionCommentReturnsParser(
+      functionCommentText
+    );
+    const functionCommentRaisesParser = new FunctionCommentRaisesParser(
+      functionCommentText
+    );
+
+    const parsingTarget = [
+      {
+        attributeName: "storage_var",
+        functionName: "ERC20_allowances",
+        functionSignature: {
+          implicitArgs: null,
+          explicitArgs: [
+            { name: "owner", type: "felt" },
+            { name: "spender", type: "felt" },
+          ],
+          returns: [{ name: "allowance", type: "Uint256" }],
+        },
+        functionComment: {
+          desc: [
+            {
+              name: "",
+              type: "",
+              desc: "Store the amount of tokens that an owner is allowed to delegate to a spender",
+              charIndex: { start: 22, end: 98 },
+            },
+          ],
+          implicitArgs: null,
+          explicitArgs: [
+            {
+              name: "owner",
+              type: "felt",
+              desc: "The address of the owner",
+              charIndex: { start: 128, end: 165 },
+            },
+            {
+              name: "spender",
+              type: "felt",
+              desc: "The address of the spender",
+              charIndex: { start: 174, end: 215 },
+            },
+          ],
+          returns: [
+            {
+              name: "allowance",
+              type: "Uint256",
+              desc: "The amount of tokens that an owner is allowed to delegate to a spender",
+              charIndex: { start: 239, end: 329 },
+            },
+          ],
+          raises: null,
+          charIndex: { start: 2461, end: 2790 },
+        },
+      },
+    ];
+
+    var parsingOutput = [
+      {
+        attributeName: functionSignatureParser.getAttributeName(
+          functionScopes![scopeNumber].text
+        ),
+        functionName: functionSignatureParser.getFunctionName(
+          functionScopes![scopeNumber].text
+        ),
+        functionSignature: {
+          implicitArgs: functionSignatureParser.getImplicitArgs(
+            functionScopes![scopeNumber].text
+          ),
+          explicitArgs: functionSignatureParser.getExplicitArgs(
+            functionScopes![scopeNumber].text
+          ),
+          returns: functionSignatureParser.getReturns(
+            functionScopes![scopeNumber].text
+          ),
+        },
+        functionComment: {
+          desc: functionCommentDescParser.parseCommentLines(
+            functionCommentScope!.text
+          ),
+          implicitArgs: functionCommentImplicitArgsParser.parseCommentLines(
+            functionCommentScope!.text
+          ),
+          explicitArgs: functionCommentExplicitArgsParser.parseCommentLines(
+            functionCommentScope!.text
+          ),
+          returns: functionCommentReturnsParser.parseCommentLines(
+            functionCommentScope!.text
+          ),
+          raises: functionCommentRaisesParser.parseCommentLines(
+            functionCommentScope!.text
+          ),
+          charIndex: {
+            start: functionCommentScope!.start,
+            end: functionCommentScope!.end,
+          },
+        },
+      },
+    ];
+
+    assert.deepEqual(parsingTarget, parsingOutput, "failed to parse");
+    const { charIndex, ...parsingOutputWithoutCharIndex } =
+      parsingOutput[0].functionComment;
+
+    var commentParsingResult = [];
+
+    for (let [key, values] of Object.entries(parsingOutputWithoutCharIndex)) {
+      if (values) {
+        for (const value of values) {
+          const charIndex = value.charIndex;
+          var char = "";
+          for (
+            let i = functionCommentScope!.start + charIndex.start;
+            i < functionCommentScope!.start + charIndex.end;
+            i++
+          ) {
+            char += text.at(i);
+          }
+          const commentParsing = {
+            [key]: char,
+          };
+          commentParsingResult.push(commentParsing);
+        }
+      }
+    }
+    const textTarget = [
+      {
+        desc: "Store the amount of tokens that an owner is allowed to delegate to a spender",
+      },
+      { explicitArgs: "owner(felt): The address of the owner" },
+      { explicitArgs: "spender(felt): The address of the spender" },
+      {
+        returns:
+          "allowance(Uint256): The amount of tokens that an owner is allowed to delegate to a spender",
+      },
+    ];
+    assert.deepEqual(textTarget, commentParsingResult, "failed to parse");
+
+    var functionCommentParsingResult = "";
+    for (var i = charIndex!.start; i < charIndex!.end; i++) {
+      functionCommentParsingResult += text.at(i);
+    }
+
+    const functionCommentTarget = `
+    # Desc: 
+    #   Store the amount of tokens that an owner is allowed to delegate to a spender
+    # Explicit args:
+    #   owner(felt): The address of the owner
+    #   spender(felt): The address of the spender
+    # Returns:
+    #   allowance(Uint256): The amount of tokens that an owner is allowed to delegate to a spender`;
 
     assert.equal(
       functionCommentTarget,
